requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=2.0.0  # Apache-2.0
PyYAML>=3.10.0  # MIT
six>=1.9.0  # MIT
Jinja2>=2.8,!=2.9.0,!=2.9.1,!=2.9.2,!=2.9.3,!=2.9.4  # BSD License (3 clause)
yaql>=1.1.0  # Apache 2.0 License
